pyproject.toml
[project]
name = "viz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "foxglove-schemas-protobuf>=0.3.0",
    "foxglove-websocket>=0.1.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pyserial>=3.5",
]
